pyproject.toml
[tool.poetry]
name = "rest-serializer-to-typescript"
version = "0.1.6"
description = ""
authors = ["Abdollah Keshtkar <hamadkeshtkar@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
